+++ src/ext_widgets/scroll_view/list_view_h.c
@@ -129,7 +129,7 @@ static ret_t list_view_h_on_scroll_view_layout_children(widget_t* widget) {
       widget_t* iter = (widget_t*)darray_get(widget->children, i);
 
       if (widget->children_layout == NULL) {
-        widget_move_resize(iter, x, y, w, h);
+        widget_move_resize_ex(iter, x, y, w, h, FALSE);
         widget_layout(iter);
       }
       x = iter->x + iter->w + spacing;