+++ src/ext_widgets/scroll_view/scroll_view.c
@@ -563,10 +563,10 @@ static ret_t scroll_view_get_prop(widget_t* widget, const char* name, value_t* v
   scroll_view_t* scroll_view = SCROLL_VIEW(widget);
   return_value_if_fail(scroll_view != NULL && name != NULL && v != NULL, RET_BAD_PARAMS);
 
-  if (tk_str_eq(name, WIDGET_PROP_VIRTUAL_W) || tk_str_eq(name, WIDGET_PROP_LAYOUT_W)) {
+  if (tk_str_eq(name, WIDGET_PROP_VIRTUAL_W)) {
     value_set_int(v, tk_max(widget->w, scroll_view->virtual_w));
     return RET_OK;
-  } else if (tk_str_eq(name, WIDGET_PROP_VIRTUAL_H) || tk_str_eq(name, WIDGET_PROP_LAYOUT_H)) {
+  } else if (tk_str_eq(name, WIDGET_PROP_VIRTUAL_H)) {
     value_set_int(v, tk_max(widget->h, scroll_view->virtual_h));
     return RET_OK;
   } else if (tk_str_eq(name, WIDGET_PROP_XOFFSET)) {